moved list image view to views/Image in preperation of improving add-image view
