+++ archive/views/__init__.py
@@ -6,3 +6,4 @@ from .person import *
 from .tag import *
 from .usermanagement import *
 from .tree import *
+from .Image.image_list import *
